# services/parking-service/openapi.json
{"openapi":"3.1.0","info":{"title":"Parking Service","description":"Parking Service Microservice for Parking Management System","version":"2.0.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["root"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/parking-lots/":{"post":{"tags":["v1","parking-lots"],"summary":"Create Parking Lot","operationId":"create_parking_lot_v1_parking_lots__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingLotCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingLotResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["v1","parking-lots"],"summary":"Get Parking Lots","operationId":"get_parking_lots_v1_parking_lots__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ParkingLotResponse"},"title":"Response Get Parking Lots V1 Parking Lots  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parking-lots/{lot_id}":{"get":{"tags":["v1","parking-lots"],"summary":"Get Parking Lot","operationId":"get_parking_lot_v1_parking_lots__lot_id__get","parameters":[{"name":"lot_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Lot Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingLotResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["v1","parking-lots"],"summary":"Update Parking Lot","operationId":"update_parking_lot_v1_parking_lots__lot_id__put","parameters":[{"name":"lot_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Lot Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingLotUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingLotResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["v1","parking-lots"],"summary":"Delete Parking Lot","operationId":"delete_parking_lot_v1_parking_lots__lot_id__delete","parameters":[{"name":"lot_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Lot Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parking-lots/{lot_id}/availability":{"get":{"tags":["v1","parking-lots"],"summary":"Get Availability","operationId":"get_availability_v1_parking_lots__lot_id__availability_get","parameters":[{"name":"lot_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Lot Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parking-lots/{lot_id}/reserve":{"post":{"tags":["v1","parking-lots"],"summary":"Reserve Spot","operationId":"reserve_spot_v1_parking_lots__lot_id__reserve_post","parameters":[{"name":"lot_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Lot Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parking-lots/{lot_id}/release":{"post":{"tags":["v1","parking-lots"],"summary":"Release Spot","operationId":"release_spot_v1_parking_lots__lot_id__release_post","parameters":[{"name":"lot_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Lot Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/auth/register":{"post":{"tags":["v1","auth"],"summary":"Register","description":"Register a new user","operationId":"register_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/auth/login":{"post":{"tags":["v1","auth"],"summary":"Login","description":"Login user","operationId":"login_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/auth/me":{"get":{"tags":["v1","auth"],"summary":"Get Current User Info","description":"Get current user info","operationId":"get_current_user_info_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/v1/auth/protected":{"get":{"tags":["v1","auth"],"summary":"Protected Route","description":"Protected route example","operationId":"protected_route_v1_auth_protected_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/v1/health":{"get":{"tags":["v1"],"summary":"Health Check","operationId":"health_check_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/ping":{"get":{"tags":["v1"],"summary":"Ping","operationId":"ping_v1_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/":{"get":{"tags":["v1"],"summary":"Root","operationId":"root_v1__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ParkingLotCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"address":{"type":"object","title":"Address"},"location":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Location"},"total_spots":{"type":"integer","title":"Total Spots"},"price_per_hour":{"type":"number","title":"Price Per Hour"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type","default":"standard"},"amenities":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Amenities"},"features":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Features"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"}},"type":"object","required":["name","address","total_spots","price_per_hour"],"title":"ParkingLotCreate"},"ParkingLotResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"address":{"type":"object","title":"Address"},"location":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Location"},"total_spots":{"type":"integer","title":"Total Spots"},"available_spots":{"type":"integer","title":"Available Spots"},"reserved_spots":{"type":"integer","title":"Reserved Spots"},"price_per_hour":{"type":"number","title":"Price Per Hour"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"amenities":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Amenities"},"features":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Features"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"review_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Review Count"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","name","description","address","location","total_spots","available_spots","reserved_spots","price_per_hour","type","status","amenities","features","phone","email","website","rating","review_count","created_at","updated_at"],"title":"ParkingLotResponse"},"ParkingLotUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"address":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Address"},"location":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Location"},"total_spots":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Spots"},"price_per_hour":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Per Hour"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"amenities":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Amenities"},"features":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Features"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"}},"type":"object","title":"ParkingLotUpdate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"},"full_name":{"type":"string","title":"Full Name"}},"type":"object","required":["username","email","password","full_name"],"title":"UserCreate"},"UserLogin":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserLogin"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}